Refactor user-service entrypoint and document API
